requirements.txt
Flask
mysql-connector-python
python-dotenv
requests
beautifulsoup4
SQLAlchemy
pg8000